Instructions for new conf
There is an entry that changes in the configuration file compared to the previous version. Creating a script to make changes to the old file would make little sense, because you should add as many lines as there are any changes between the various versions and take into account the version from which you start to update. In this case, it is easier to add an instruction.